Tarea-Semana-12/requirements.txt
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mpi4py==3.1.5
nltk==3.8.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
regex==2023.10.3
six==1.16.0
tqdm==4.66.1
